backend/requirements.txt
beautifulsoup4==4.12.2
black==25.1.0
django-cors-headers==4.4.0
django-debug-toolbar==4.4.6
djangorestframework==3.14.0
flake8==7.2.0
isort==6.0.1
pipdeptree==2.26.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.0
requests==2.31.0